// src/gen/models/apiResponseVoid.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * STORE ME API
 * Hệ thống lưu trữ file cho phép người dùng upload, download và quản lý file của mình.

 * OpenAPI spec version: 1.0 beta
 */
import type { ErrorVO } from './errorVO';

export interface ApiResponseVoid {
  statusCode?: number;
  message?: string;
  data?: unknown;
  error?: ErrorVO;
}